Refactor cache population and number plate generation: Change PopulateCache function to accept uint8; update main to populate cache with 10 drivers; modify number plate generation to use a fixed letter prefix; enhance JSON encoding in handler responses.

+++ web_server/lib.go
@@ -33,8 +33,8 @@ func (d *Drivers) GenData() {
 	}
 }
 
-func PopulateCache(n int) {
-	for i := 0; i < n; i++ {
+func PopulateCache(n uint8) {
+	for i := 0; i < int(n); i++ {
 		var driver Drivers
 		driver.GenData()
